Вынес компонент ссылки, так как используется неоднократно, создал страницу-заготовку для форума
