+++ packages/client/src/assets/constants/colors.ts
@@ -1,0 +1,6 @@
+export enum Colors {
+  yellow = 'yellow',
+  blue = 'blue',
+  black = 'black',
+}
+export type TColorsType = Colors.blue | Colors.black | Colors.yellow;
