+++ packages/client/src/components/ErrorPageConstructor/ErrorPageConstructor.tsx
@@ -1,19 +1,14 @@
 import styles from './styles.module.scss';
-import Link from '../Link';
-
-enum Colors {
-  yellow = 'yellow',
-  blue = 'blue',
-  black = 'black',
-}
+import { Colors, TColorsType } from '../../assets/constants/colors';
+import CustomLink from '../CustomLink/CustomLink';
 
 type TErrorPageConstructor = {
   image: string;
   title?: string;
   linkText: string;
   linkPath: string;
-  linkColor: Colors.blue | Colors.black | Colors.yellow;
-  backgroundColor?: Colors.black | Colors.yellow | Colors.blue;
+  linkColor: TColorsType;
+  backgroundColor?: TColorsType;
 };
 
 const ErrorPageConstructor: React.FC<TErrorPageConstructor> = ({
@@ -35,25 +30,17 @@ const ErrorPageConstructor: React.FC<TErrorPageConstructor> = ({
     }
   };
 
-  const linkTextClassname = () => {
-    switch (linkColor) {
-      case Colors.yellow:
-        return styles.linkTextYellow;
-      case Colors.blue:
-        return styles.linkTextBlue;
-      default:
-        return styles.linkTextBlack;
-    }
-  };
   return (
     <section className={sectionClassname()}>
       <div className={styles.imageContainer}>
         <img className={styles.image} src={image} alt="image" />
         {title ? <h3 className={styles.title}>{title}</h3> : null}
       </div>
-      <Link to={linkPath}>
-        <p className={linkTextClassname()}>{linkText}</p>
-      </Link>
+      <CustomLink
+        linkPath={linkPath}
+        linkText={linkText}
+        linkColor={linkColor}
+      />
     </section>
   );
 };
