+++ packages/client/src/pages/404/NotFoundPage.tsx
@@ -1,13 +1,7 @@
 import ErrorPageConstructor from '../../components/ErrorPageConstructor/ErrorPageConstructor';
 import notFoundImage from '../../assets/images/404-image.svg';
 import { ROUTES } from '../../constants/routes';
-
-enum Colors {
-  yellow = 'yellow',
-  blue = 'blue',
-  black = 'black',
-}
-
+import { Colors } from '../../assets/constants/colors';
 const NotFoundPage = () => {
   const title = 'Not found';
   const linkText = 'back to the game';
