+++ packages/client/src/pages/Forum/Forum.tsx
@@ -1,0 +1,7 @@
+import styles from './styles.module.scss';
+
+const Forum = () => {
+  return <div className={styles.container}>Forum</div>;
+};
+
+export default Forum;
