+++ packages/client/src/pages/Forum/styles.module.scss
@@ -1,0 +1,8 @@
+@import '../../assets/styles/mixins';
+@import '../../assets/styles/variables';
+
+.container {
+  background: url("../../assets/images/forum-background.svg") no-repeat right bottom;
+  width: 100%;
+  height: 100%;
+}
