+++ packages/client/src/routes/Main/Main.tsx
@@ -6,6 +6,7 @@ import GameOver from '../../pages/GameOver/GameOver';
 import { ROUTES } from '../../constants/routes';
 import Profile from '../../pages/Profile/Profile';
 import Leaderboard from '../../pages/Leaderboard/Leaderboard';
+import Forum from '../../pages/Forum/Forum';
 
 export const Main = () => {
   return (
@@ -16,7 +17,7 @@ export const Main = () => {
           <Route path={ROUTES.main.profile} element={<Profile />} />
           <Route path={ROUTES.main.root} element={<div>Game</div>} />
           <Route path={ROUTES.main.gameOver} element={<GameOver />} />
-          <Route path={ROUTES.main.forum} element={<div>Forum</div>} />
+          <Route path={ROUTES.main.forum} element={<Forum />} />
           <Route path={ROUTES.main.lead} element={<Leaderboard />} />
 
           <Route
